fix early return to fix reload issue

+++ NetworkDynamicsInspector/src/timeseries.jl
@@ -52,38 +52,42 @@ function timeseries_cards(app, session)
         known_tsplots = collect(keys(cache))
         current_tsplots = collect(values(_tsplots))
 
-        # do nothing if the TSPlots objects themselve did not change
-        Set(known_tsplots) == Set(current_tsplots) && return
+        # do nothing if the TSPlots objects themselve did not change unless on update!
+        if !isdefined(container, 0) || Set(known_tsplots) != Set(current_tsplots)
+            # del_tsplots = setdiff(known_tsplots, current_tsplots)
+            # new_tsplots = setdiff(current_tsplots, known_tsplots)
+            # FIXME: it is not possible to redisplay the same tsplot, recreate all
+            del_tsplots = known_tsplots
+            new_tsplots = current_tsplots
 
-        # del_tsplots = setdiff(known_tsplots, current_tsplots)
-        # new_tsplots = setdiff(current_tsplots, known_tsplots)
-        # FIXME: it is not possible to redisplay the same tsplot, recreate all
-        del_tsplots = known_tsplots
-        new_tsplots = current_tsplots
+            for del in del_tsplots
+                (; card, observerfunctions, plotqueue) = cache[del]
+                close(plotqueue) # close plotqueue
+                Observables.off.(observerfunctions) # disable all observer functions
+                delete!(cache, del) # delete from cache
+            end
+            for new in new_tsplots
+                key = only([key for (key, value) in _tsplots if value == new])
+                ntup = timeseries_card(app, key, session)
+                cache[new] = ntup
+            end
 
-        for del in del_tsplots
-            (; card, observerfunctions, plotqueue) = cache[del]
-            close(plotqueue) # close plotqueue
-            Observables.off.(observerfunctions) # disable all observer functions
-            delete!(cache, del) # delete from cache
+            cards = [cache[ts].card for ts in values(_tsplots)]
+            # update the display by notifying the content
+            container[] = DOM.div(cards; class="timeseries-stack")
         end
-        for new in new_tsplots
-            key = only([key for (key, value) in _tsplots if value == new])
-            ntup = timeseries_card(app, key, session)
-            cache[new] = ntup
-        end
-
-        cards = [cache[ts].card for ts in values(_tsplots)]
-        # update the display by notifying the content
-        container[] = DOM.div(cards; class="timeseries-stack")
 
         nothing
     end
 
     # update the selected components in the graphplot when the active tsplot changes
     on(app.active_tsplot; update=true) do active
-        activesel = app.tsplots[][active].selcomp[]
-        app.graphplot._selcomp[] = activesel
+        if haskey(app.tsplots[], active)
+            activesel = app.tsplots[][active].selcomp[]
+            app.graphplot._selcomp[] = activesel
+        else
+            app.active_tsplot[] = first(keys(app.tsplots[]))
+        end
     end
 
     return container
